SupportAgent/requirements.txt
Flask
scikit-learn
python-dotenv
requests
numpy
tensorflow
matplotlib

google-generativeai

transformers 
scikit-learn
numpy
